Update Backing up Data to DreamObjects Using Duplicity.html
Adding instructions no how to restore data (since I had to Google it and that seems a tad ridiculous lol). Sorry if I missed a closing div tag or if my formatting is super janky. 

I would also like to propose renaming the article to "How to Use Duplicity with DreamObjects" to keep it in line with the other application guides we have.