Update How to change the size of your DreamCompute server.html
The concept of changing size and flavor and the disconnectedness between that and what's actually being done (increasing the resource limit for an instance) has always bugged me, so I finally did something about it! lol

Oh, also need the title changed. 

How To Change The Resource Limits Of A DreamCompute Instance